requirements.txt
# Core dependencies
requests>=2.28.0
gspread>=5.0.0
google-api-python-client>=2.88.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0